+++ lib/motion-aasm.rb
@@ -16,6 +16,7 @@ exclude_files = [
   'aasm/persistence/active_record_persistence.rb',
   'aasm/persistence/dynamoid_persistence.rb',
   'aasm/persistence/mongoid_persistence.rb',
+  'aasm/persistence/no_brainer_persistence.rb',
   'aasm/persistence/sequel_persistence.rb',
   'aasm/persistence/redis_persistence.rb'
 ]